// src/types.ts
export type ConversationResponseEvent = {
    message?: Message
    conversation_id?: string
    error?: string | null
}

export type Message = {
    id: string
    content: MessageContent
    role: string
    user: string | null
    create_time: string | null
    update_time: string | null
    end_turn: null
    weight: number
    recipient: string
    metadata: MessageMetadata
}

export type MessageContent = {
    content_type: string
    parts: string[]
}

export type MessageMetadata = any
export type MessageActionType = 'next' | 'variant'

export type SendMessageOptions = {
    conversationId?: string
    parentMessageId?: string
    messageId?: string
    action?: MessageActionType
    timeoutMs?: number
    onProgress?: (partialResponse: string) => void
    onConversationResponse?: (response: ConversationResponseEvent) => void
    abortSignal?: AbortSignal
}

export type SendConversationMessageOptions = Omit<
    SendMessageOptions,
    'conversationId' | 'parentMessageId'
    >